pyproject.toml
[project]
name = "zacky-ia"
version = "0.1.0"
description = "AI-powered support assistant for Zendesk"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # API & Web
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "streamlit>=1.31.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Data & ML
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.5",

    # OpenAI
    "openai>=1.12.0",

    # Google Cloud
    "google-cloud-bigquery>=3.17.0",
    "google-cloud-storage>=2.14.0",
    "db-dtypes>=1.2.0",

    # Zendesk
    "zenpy>=2.0.41",

    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
